requirements.txt
gradio~=3.0.21
numpy~=1.23.0
h5py~=3.7.0
opencv-python~=4.6.0.66
pyinflect~=0.5.1
tensorflow~=2.9.1
keras~=2.9.0
nltk~=3.7
scikit-learn~=1.1.1
pandas~=1.4.3
transformers~=4.20.1
torch~=1.12.0
Pillow~=9.2.0